extract some more tests
