+++ qcodes/tests/parameter/test_parameter.py
@@ -2,13 +2,9 @@
 Test suite for parameter
 """
 from unittest import TestCase
-import pytest
 
-from qcodes.instrument.parameter import Parameter, _BaseParameter
+from qcodes.instrument.parameter import Parameter
 from qcodes.tests.instrument_mocks import DummyInstrument
-
-from .conftest import (OverwriteGetParam, OverwriteSetParam,
-                       GetSetRawParameter)
 
 
 class TestStandardParam(TestCase):
@@ -29,7 +25,7 @@ class TestStandardParam(TestCase):
 
     def test_param_cmd_with_parsing(self):
         p = Parameter('p_int', get_cmd=self.get_p, get_parser=int,
-                              set_cmd=self.set_p, set_parser=self.parse_set_p)
+                      set_cmd=self.set_p, set_parser=self.parse_set_p)
 
         p(5)
         self.assertEqual(self._p, '5')
@@ -101,97 +97,4 @@ class TestManualParameterValMapping(TestCase):
         assert self.instrument.myparameter.raw_value == 0
 
 
-def test_parameter_with_overwritten_get_raises():
-    """
-    Test that creating a parameter that overwrites get and set raises runtime errors
-    """
 
-    with pytest.raises(RuntimeError) as record:
-        a = OverwriteGetParam(name='foo')
-    assert "Overwriting get in a subclass of _BaseParameter: foo is not allowed." == str(record.value)
-
-
-def test_parameter_with_overwritten_set_raises():
-    """
-    Test that creating a parameter that overwrites get and set raises runtime errors
-    """
-    with pytest.raises(RuntimeError) as record:
-        a = OverwriteSetParam(name='foo')
-    assert "Overwriting set in a subclass of _BaseParameter: foo is not allowed." == str(record.value)
-
-
-def test_unknown_args_to_baseparameter_warns():
-    """
-    Passing an unknown kwarg to _BaseParameter should trigger a warning
-    """
-    with pytest.warns(Warning):
-        a = _BaseParameter(name='Foo',
-                           instrument=None,
-                           snapshotable=False)
-
-
-@pytest.mark.parametrize("get_cmd, set_cmd", [(False, False), (False, None), (None, None), (None, False),
-                                              (lambda: 1, lambda x: x)])
-def test_gettable_settable_attributes_with_get_set_cmd(get_cmd, set_cmd):
-    a = Parameter(name='foo',
-                  get_cmd=get_cmd,
-                  set_cmd=set_cmd)
-    expected_gettable = get_cmd is not False
-    expected_settable = set_cmd is not False
-
-    assert a.gettable is expected_gettable
-    assert a.settable is expected_settable
-
-
-@pytest.mark.parametrize("baseclass", [_BaseParameter, Parameter])
-def test_gettable_settable_attributes_with_get_set_raw(baseclass):
-    """Test that parameters that have get_raw,set_raw are
-    listed as gettable/settable and reverse."""
-
-    class GetSetParam(baseclass):
-        def __init__(self, *args, initial_value=None, **kwargs):
-            self._value = initial_value
-            super().__init__(*args, **kwargs)
-
-        def get_raw(self):
-            return self._value
-
-        def set_raw(self, value):
-            self._value = value
-
-    a = GetSetParam('foo', instrument=None, initial_value=1)
-
-    assert a.gettable is True
-    assert a.settable is True
-
-    b = _BaseParameter('foo', None)
-
-    assert b.gettable is False
-    assert b.settable is False
-
-
-@pytest.mark.parametrize("working_get_cmd", (False, None))
-@pytest.mark.parametrize("working_set_cmd", (False, None))
-def test_get_raw_and_get_cmd_raises(working_get_cmd, working_set_cmd):
-    with pytest.raises(TypeError, match="get_raw"):
-        GetSetRawParameter(name="param1", get_cmd="GiveMeTheValue", set_cmd=working_set_cmd)
-    with pytest.raises(TypeError, match="set_raw"):
-        GetSetRawParameter(name="param2", set_cmd="HereIsTheValue {}", get_cmd=working_get_cmd)
-    GetSetRawParameter("param3", get_cmd=working_get_cmd, set_cmd=working_set_cmd)
-
-
-def test_get_on_parameter_marked_as_non_gettable_raises():
-    a = Parameter("param")
-    a._gettable = False
-    with pytest.raises(TypeError, match="Trying to get a parameter that is not gettable."):
-        a.get()
-
-
-def test_set_on_parameter_marked_as_non_settable_raises():
-    a = Parameter("param", set_cmd=None)
-    a.set(2)
-    assert a.get() == 2
-    a._settable = False
-    with pytest.raises(TypeError, match="Trying to set a parameter that is not settable."):
-        a.set(1)
-    assert a.get() == 2
